styling the rest of the pages
